Rename connector module to protocols (#393)
