+++ docs/concepts/index.md
@@ -3,23 +3,23 @@
 
 ProxyStore is composed of three main components: the
 [`Proxy`][proxystore.proxy.Proxy],
-[`Connector`][proxystore.connectors.connector.Connector],
+[`Connector`][proxystore.connectors.protocols.Connector],
 and [`Store`][proxystore.store.base.Store].
 
 The [`Proxy`][proxystore.proxy.Proxy] model provides *pass-by-reference*
 semantics and *just-in-time* object resolution transparently to consumers.
 
-The [`Connector`][proxystore.connectors.connector.Connector] is a
+The [`Connector`][proxystore.connectors.protocols.Connector] is a
 [`Protocol`][typing.Protocol] that defines the low-level
 interface to a mediated communication channel or object store.
-Many [`Connector`][proxystore.connectors.connector.Connector] implementations
+Many [`Connector`][proxystore.connectors.protocols.Connector] implementations
 are provided in the [`proxystore.connectors`][proxystore.connectors] module,
 and users can easily create their own.
 
 The [`Store`][proxystore.store.base.Store] is a high-level abstraction of an
 object store and the intended means by which an application uses ProxyStore.
 The [`Store`][proxystore.store.base.Store] is initialized with
-a [`Connector`][proxystore.connectors.connector.Connector] and provides
+a [`Connector`][proxystore.connectors.protocols.Connector] and provides
 extra functionality like caching and serialization. Most important is that the
 [`.proxy()`][proxystore.store.base.Store.proxy] method is provided which can
 produce a [`Proxy`][proxystore.proxy.Proxy] of an arbitrary object put in the
